Added contact_info approval column to user table. Added Hobbies and interests. Replaced profile view with three resume styles of student, employed, and independent. Added phone and Linked-In contacts to profile table.
Added contact_info approval column to user table which must be set to 1 in order to make profiles with contact information. If contact_info is set, the contacts.php page follows save for the edit screen.

Added phone and Linked-In contacts to profile table. Added new methods to util.php named changeResumeStyle, insertPhone, insertLinkedin.

Added hobbies and interests which affected add.php, edit.php, and view.php. For index.php, replaced profile view file view.php, Replaced view.php with three resume styles of student.php, employed.php, and independent.php. Radio button added to add and edit pages to select resume style.

The Hobby and HobbyList tables are very similar to the setup for the Skill and SkillSet tables for job skills. Similar to the SkillSet table, the HobbyList table is a one-to-one relationship table between the Profile and Hobby table.

The JavaScript file script.js was changed for the doValidate functions which run for submitting data. The JSON language check for the last text box that was used is dropped, primarily because it delays the final submission process to much; there were warning messages concerning the XMLHttpRequest Standard.

There always seem to be many css changes that affect the layouts.