Correct profile.php by removing the deletions that are now included in contacts.php and editInterestsActivities.php. Minor CSS changes for white space and links.
The basic process always removes current database entries for a profile and replaces them with a new set. These operations must be contained in a single editing file. The skill operations are within the 'editSkill' file, the basic profile information is within the 'editProfile' file and so on. Previously, there was a single edit file that contained everything which now only handles the basic profile information like name and e-mail address.

Multiple CSS styles had been present for links including a style inherited from a parent paragraph tag.  A lighter blue was chosen for links instead of the default deep blue.

1.5-em units seems to be an adequate space between sections compared to 2-em units.